+++ fr.rostren.tracker.ui/src/fr/rostren/tracker/ui/properties/wizards/AddAccountWizard.java
@@ -84,7 +84,7 @@ public class AddAccountWizard extends Wizard {
 	 * Returns the account amount
 	 * @return the account amount
 	 */
-	public float getAccountAmount() {
+	public double getAccountAmount() {
 		return page.getAccountAmount();
 	}
 }
